Fix for macos-specific issues around transforming xcresult files in the quarantine command.
