+++ cli/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod api_client;
+mod junit_utils;
 pub mod quarantine;
 pub mod runner;
 pub mod scanner;
